+++ includes/backend.php
@@ -21,29 +21,54 @@ defined( 'ABSPATH' ) or exit;
  * @since 0.0.3
  */
 
-function plugin_backend_scripts()
-{
-    $currentScreen = get_current_screen();
+function plugin_backend_scripts() {
+    $current_screen = get_current_screen();
 
-    if( 'upload' === $currentScreen->id ) :
+    if( ( 'upload' === $current_screen->id ) or ( 'attachment' === $current_screen->id ) ):
 
-        global $mode;
+        /**
+         * Enqueue style settings for both upload and attachment page
+         */
 
-        if( 'grid' === $mode ) :
+        wp_enqueue_style(
+            'mdb_license_management-backend-style',
+            PLUGIN_URL . 'assets/build/css/backend.min.css',
+            [],
+            PLUGIN_VERSION
+        );
 
-            wp_enqueue_style(
-                'mdb_license_management-backend-style',
-                PLUGIN_URL . 'assets/css/admin.css'
-            );
 
+        /**
+         * Enqueue script for upload page only
+         */
+
+        if( 'upload' === $current_screen->id ) :
             wp_enqueue_script(
-                'mdb_license_management-backend-script',
-                PLUGIN_URL . 'assets/js/admin.js',
-                'jquery',
+                'mdb_license_management-backend-upload-script',
+                PLUGIN_URL . 'assets/build/js/backend-upload.min.js',
+                [
+                    'jquery'
+                ],
                 PLUGIN_VERSION,
                 true
             );
+        endif;
 
+
+        /**
+         * Enqueue script for attachment page only
+         */
+
+        if( 'attachment' === $current_screen->id ) :
+            wp_enqueue_script(
+                'mdb_license_management-backend-attachment-script',
+                PLUGIN_URL . 'assets/build/js/backend-attachment.min.js',
+                [
+                    'jquery'
+                ],
+                PLUGIN_VERSION,
+                true
+            );
         endif;
 
     endif;