+++ includes/classes/index.php
@@ -1,7 +1,0 @@
-<?php
-/** Prevent direct access */
-
-defined( 'ABSPATH' ) or exit;
-
-
-require_once( 'class-main-table.php' );