+++ includes/core/index.php
@@ -1,9 +1,0 @@
-<?php
-/** Prevent direct access */
-
-defined( 'ABSPATH' ) or exit;
-
-
-require_once( 'core-indexing.php' );
-require_once( 'core-media-record.php' );
-require_once( 'core-media-library.php' );