+++ plugin.php
@@ -6,7 +6,7 @@
  * Author:          Marco Di Bella
  * Author URI:      https://www.marcodibella.de
  * License:         MIT License
- * Version:         0.0.4
+ * Version:         1.0.0
  * Text Domain:     mdb-license-management
  *
  * @author  Marco Di Bella
@@ -24,28 +24,25 @@ defined( 'ABSPATH' ) or exit;
 
 /** Variables and definitions */
 
-define( __NAMESPACE__ . '\PLUGIN_VERSION', '0.0.4' );
+define( __NAMESPACE__ . '\PLUGIN_VERSION', '1.0.0' );
 define( __NAMESPACE__ . '\PLUGIN_DIR', plugin_dir_path( __FILE__ ) );
 define( __NAMESPACE__ . '\PLUGIN_URL', plugin_dir_url( __FILE__ ) );
 
-define( 'MEDIA_STATE_UNKNOWN', 0 );
-define( 'MEDIA_STATE_NO_CREDIT', 1 );
-define( 'MEDIA_STATE_SIMPLE_CREDIT', 2 );
-define( 'MEDIA_STATE_LICENSED', 3 );
 
-const table_media    = 'mdb_lv_media';
-const table_licenses = 'mdb_lv_licenses';
-
-define( __NAMESPACE__ . '\MENU_SLUG', 'mdb_license_management' );
-
+const TABLE_LICENSES = 'mdb_lv_licenses';
+const TABLE_MEDIA    = 'mdb_lv_media';
 
 
 /** Include files */
 
-require_once( PLUGIN_DIR . 'includes/classes/index.php' );
-require_once( PLUGIN_DIR . 'includes/core/index.php' );
+require_once( PLUGIN_DIR . 'includes/definitions/media-states.php' );
+require_once( PLUGIN_DIR . 'includes/definitions/licenses.php' );
+
+require_once( PLUGIN_DIR . 'includes/classes/class-media-record.php' );
+
+require_once( PLUGIN_DIR . 'includes/media-library/upload.php' );
+require_once( PLUGIN_DIR . 'includes/media-library/attachment.php' );
 
 require_once( PLUGIN_DIR . 'includes/theme-integration.php' );
-require_once( PLUGIN_DIR . 'includes/mainpage.php' );
 require_once( PLUGIN_DIR . 'includes/backend.php' );
 require_once( PLUGIN_DIR . 'includes/setup.php' );