+++ CHANGELOG.md
@@ -4,6 +4,22 @@
 
 <br>
 
+### 1.0.0
+Released: 06-08-2023
+
+* New: Add minified scripts and styles
+* New: Add class MediaRecord
+* New: Add support for Pixabay, Pexels and Unsplash license
+* Improve: Complete overhaul of the plugin files and structure
+* Improve: Update German translation
+* Improve: Replace license table with constant array
+* Improve: Add more sanitizing/escaping
+* Improve: Remove media indexing
+* Improve: Remove admin menu
+* Improve: Remove listing of all available licenses
+* Bugfix: Fix scripting in attachment pages
+
+
 ### 0.0.4
 Released: 03-08-2023